requirements.txt
# Project dependencies for Aroma College Chatbot (using Ollama local models)
langchain>=0.0.267
langchain-community>=0.0.15
python-dotenv>=1.0.0
chromadb>=0.4.18
pynvml>=11.5.0
pandas>=2.0.0
pyPDF>=3.17.1
docx2txt>=0.8
streamlit>=1.29.0
faiss-cpu>=1.7.4
hf-transfer>=0.1.4
sentence-transformers>=2.2.2
ollama>=0.1.5
torch>=2.0.0
